requirements.txt
flask
flask_limiter
requests
pillow
face_recognition
